requirements.txt
streamlit>=1.32.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
pypdf>=4.1.0
streamlit-drawable-canvas>=0.9.3
